Food incorporated/Kitchen_app/requirements.txt
flask
flask-cors
flask-sqlalchemy
python-dotenv
gunicorn
requests
psycopg2-binary